+++ kernel/src/HAL/drivers/ACPI/SDT.cpp
@@ -2,11 +2,10 @@
 
 bool doChecksum(ACPISDTHeader* h) {
     unsigned char sum = 0;
- 
-    for (int i = 0; i < h->Length; i++)
-    {
+
+    for (uint32_t i = 0; i < h->Length; i++) {
         sum += ((char*)h)[i];
     }
- 
+
     return sum == 0;
 }