+++ kernel/src/HAL/drivers/ACPI/XSDT.cpp
@@ -18,3 +18,7 @@ ACPISDTHeader* getOtherSDT(uint64_t index) {
     uint64_t* start = (uint64_t*)((uint64_t)g_XSDT + sizeof(ACPISDTHeader));
     return (ACPISDTHeader*)(start[index]);
 }
+
+uint64_t getSDTCount() {
+    return (g_XSDT->Length - sizeof(ACPISDTHeader)) / sizeof(uint64_t);
+}