+++ kernel/src/HAL/drivers/Device.hpp
@@ -1,0 +1,13 @@
+#ifndef _HAL_DEVICE_HPP
+#define _HAL_DEVICE_HPP
+
+class Device {
+public:
+    Device();
+    virtual ~Device();
+
+    virtual const char* getVendorName();
+    virtual const char* getDeviceName();
+};
+
+#endif /* _HAL_DEVICE_HPP */